% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{noaa_app}
\alias{noaa_app}
\title{Creates a NOAA Shiny App}
\usage{
noaa_app(selected_location = "Ed Levin, CA (1750 ft)")
}
\arguments{
\item{selected_location}{(character) ...}
}
\value{
A Shiny App.
}
\description{
Creates a NOAA Shiny App
}
