% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{noaa_url}
\alias{noaa_url}
\title{Get NOAA Weather Forcast URL for a Location}
\usage{
noaa_url(lat, lon, format = c("xml", "html"))
}
\arguments{
\item{lat, lon}{(numeric) A lattitude and the longitude.}

\item{format}{(character) Should an XML or an HTML document be retrieved?}
}
\value{
(character) The URL.
}
\description{
Get NOAA Weather Forcast URL for a Location
}
