% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{noaa_url}
\alias{noaa_url}
\alias{read_noaa}
\title{Get NOAA Weather Forcast URL for a Location}
\usage{
noaa_url(lat, lon, format = c("xml", "html"))

read_noaa(url, tz = timezone())
}
\arguments{
\item{lat, lon}{(numeric) A lattitude and the longitude.}

\item{format}{(character) Should an XML or an HTML document be retrieved?}

\item{url}{(character) A NOAA URL to download and parse.}

\item{tz}{(character) The timezone for all retrieved timestamps.}
}
\value{
`noaa_url()` returns a URL (character).

`read_noaa()` returns a [tibble::tibble] [base::data.frame] with
forecasted weather data.
}
\description{
Get NOAA Weather Forcast URL for a Location

Get NOAA Weather Forcast for a Location
}
\examples{
coord <- c(lat=37.475400, lon=-121.861300)
url <- noaa_url(coord)
print(url)
wx <- read_noaa(url)
print(wx)

}
