% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{save_noaa}
\alias{save_noaa}
\title{Save NOAA Weather Forcast to File}
\usage{
save_noaa(wx, filename = NULL, path = ".", skip = TRUE)
}
\arguments{
\item{wx}{(tibble) A NOAA Weather Forecast.}

\item{filename}{(character) The filename of the saved RDS file.
If NULL (default), then the filename is uniquely created from the GPS
coordinates and the timestamp of forecast.}

\item{path}{(character) Folder where to save forecast.}

\item{skip}{(logical) If TRUE, then an already saved forecast will not
be overwritten.}
}
\value{
(character) The pathname of the saved forecast.
If skipped, then attribute `skipped` is set to TRUE.
}
\description{
Save NOAA Weather Forcast to File
}
