% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timezone}
\alias{timezone}
\title{Get or Set Time Zone used by this Package}
\usage{
timezone(new = NULL)
}
\arguments{
\item{new}{(character) If non-NULL, new timezone to be set.}
}
\value{
(character) The current timezone set.
}
\description{
Get or Set Time Zone used by this Package
}
