% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{update_noaa}
\alias{update_noaa}
\title{Update NOAA Weather Forcast Data Base}
\usage{
update_noaa(latitude, longitude, path = ".")
}
\arguments{
\item{latitude, longitude}{(numeric) The coordinates of the location.}

\item{path}{(character) Folder where to save forecast.}
}
\value{
(character) The pathname of the saved forecast.
}
\description{
Update NOAA Weather Forcast Data Base
}
