% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa.R
\name{read_noaa}
\alias{read_noaa}
\title{Get NOAA Weather Forcast for a Location}
\usage{
read_noaa(url, tz = timezone())
}
\arguments{
\item{url}{(character) A NOAA URL to download and parse.}

\item{tz}{(character) The timezone for all retrieved timestamps.}
}
\value{
(tibble data.frame) The forcasted weather data.
}
\description{
Get NOAA Weather Forcast for a Location
}
